requirements.txt
Django == 3.0
djangorestframework == 3.11
psycopg2==2.8.0
pyotp == 2.6.0
twilio == 6.52.0
python-decouple == 3.4
pillow == 8.1.0